requirements.txt
web3==6.8.0
vyper==0.4.1
openai==0.28.1
python-dotenv==1.0.0 